restore_location_update
