+++ ast/ast.go
@@ -18,6 +18,7 @@ package ast
 import (
 	"io"
 
+	. "github.com/pingcap/parser/format"
 	"github.com/pingcap/parser/model"
 	"github.com/pingcap/parser/types"
 )
