+++ ast/functions.go
@@ -18,6 +18,7 @@ import (
 	"io"
 
 	"github.com/pingcap/errors"
+	. "github.com/pingcap/parser/format"
 	"github.com/pingcap/parser/model"
 	"github.com/pingcap/parser/types"
 )
