+++ ast/stats.go
@@ -15,6 +15,7 @@ package ast
 
 import (
 	"github.com/pingcap/errors"
+	. "github.com/pingcap/parser/format"
 	"github.com/pingcap/parser/model"
 )
