+++ format/format_test.go
@@ -16,6 +16,7 @@ package format
 import (
 	"bytes"
 	"io/ioutil"
+	"strings"
 	"testing"
 
 	. "github.com/pingcap/check"
@@ -28,6 +29,7 @@ func TestT(t *testing.T) {
 }
 
 var _ = Suite(&testFormatSuite{})
+var _ = Suite(&testRestoreCtxSuite{})
 
 type testFormatSuite struct {
 }
@@ -57,4 +59,41 @@ z
 	f = FlatFormatter(buf)
 	expect = "abc3%e x y z\n "
 	checkFormat(c, f, buf, str, expect)
+}
+
+type testRestoreCtxSuite struct {
+}
+
+func (s *testRestoreCtxSuite) TestRestoreCtx(c *C) {
+	testCases := []struct {
+		flag   RestoreFlags
+		expect string
+	}{
+		{0, "key`.'\"Word\\ str`.'\"ing\\ na`.'\"Me\\"},
+		{RestoreStringSingleQuotes, "key`.'\"Word\\ 'str`.''\"ing\\' na`.'\"Me\\"},
+		{RestoreStringDoubleQuotes, "key`.'\"Word\\ \"str`.'\"\"ing\\\" na`.'\"Me\\"},
+		{RestoreStringEscapeBackslash, "key`.'\"Word\\ str`.'\"ing\\\\ na`.'\"Me\\"},
+		{RestoreKeyWordUppercase, "KEY`.'\"WORD\\ str`.'\"ing\\ na`.'\"Me\\"},
+		{RestoreKeyWordLowercase, "key`.'\"word\\ str`.'\"ing\\ na`.'\"Me\\"},
+		{RestoreNameUppercase, "key`.'\"Word\\ str`.'\"ing\\ NA`.'\"ME\\"},
+		{RestoreNameLowercase, "key`.'\"Word\\ str`.'\"ing\\ na`.'\"me\\"},
+		{RestoreNameDoubleQuotes, "key`.'\"Word\\ str`.'\"ing\\ \"na`.'\"\"Me\\\""},
+		{RestoreNameBackQuotes, "key`.'\"Word\\ str`.'\"ing\\ `na``.'\"Me\\`"},
+		{DefaultRestoreFlags, "KEY`.'\"WORD\\ 'str`.''\"ing\\' `na``.'\"Me\\`"},
+		{RestoreStringSingleQuotes | RestoreStringDoubleQuotes, "key`.'\"Word\\ 'str`.''\"ing\\' na`.'\"Me\\"},
+		{RestoreKeyWordUppercase | RestoreKeyWordLowercase, "KEY`.'\"WORD\\ str`.'\"ing\\ na`.'\"Me\\"},
+		{RestoreNameUppercase | RestoreNameLowercase, "key`.'\"Word\\ str`.'\"ing\\ NA`.'\"ME\\"},
+		{RestoreNameDoubleQuotes | RestoreNameBackQuotes, "key`.'\"Word\\ str`.'\"ing\\ \"na`.'\"\"Me\\\""},
+	}
+	var sb strings.Builder
+	for _, testCase := range testCases {
+		sb.Reset()
+		ctx := NewRestoreCtx(testCase.flag, &sb)
+		ctx.WriteKeyWord("key`.'\"Word\\")
+		ctx.WritePlain(" ")
+		ctx.WriteString("str`.'\"ing\\")
+		ctx.WritePlain(" ")
+		ctx.WriteName("na`.'\"Me\\")
+		c.Assert(sb.String(), Equals, testCase.expect, Commentf("case: %#v", testCase))
+	}
 }
